# ENH:signal: Speed-up Cython implementation of _sosfilt