--- a/scipy/signal/_sosfilt.pyx
+++ b/scipy/signal/_sosfilt.pyx
@@ -33,17 +33,21 @@ cdef void _sosfilt_float(DTYPE_floating_t [:, ::1] sos,
     cdef Py_ssize_t n_sections = sos.shape[0]
     cdef Py_ssize_t i, n, s
     cdef DTYPE_floating_t x_new, x_cur
+    cdef DTYPE_floating_t[:, ::1] zi_slice
 
+    # jumping through a few memoryview hoops to reduce array lookups,
+    # the original version is still in the gil version below.
     for i in xrange(n_signals):
+        zi_slice = zi[i, :, :]
         for n in xrange(n_samples):
 
             x_cur = x[i, n]
 
             for s in xrange(n_sections):
-                x_new = sos[s, 0] * x_cur + zi[i, s, 0]
-                zi[i, s, 0] = (sos[s, 1] * x_cur - sos[s, 4] * x_new
-                               + zi[i, s, 1])
-                zi[i, s, 1] = sos[s, 2] * x_cur - sos[s, 5] * x_new
+                x_new = sos[s, 0] * x_cur + zi_slice[s, 0]
+                zi_slice[s, 0] = (sos[s, 1] * x_cur - sos[s, 4] * x_new
+                                  + zi_slice[s, 1])
+                zi_slice[s, 1] = sos[s, 2] * x_cur - sos[s, 5] * x_new
                 x_cur = x_new
 
             x[i, n] = x_cur