As per the discussion on #12970, `sosfilt` implementation is accelerated through some rearrangements of the loops. Before this change as shown in the issue the speed difference was occasionally hitting x2 slowdowns.

Continuing the example from the linked issue for `dt=[10, 100, 1000]`  I get the following "almost" equal performance, measured through 

```python
%timeit timeseries_lp_sos = scipy.signal.sosfilt(lp_sos, timeseries)
%timeit timeseries_lp = scipy.signal.lfilter(lp_numerator, lp_denominator, timeseries)
```

```bash
dt = 10
6.74 ms ± 108 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) # This PR
6.1 ms ± 75.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) # ba performance

dt = 100
67.1 ms ± 940 µs per loop (mean ± std. dev. of 7 runs, 10 loops each) # This PR
63.6 ms ± 932 µs per loop (mean ± std. dev. of 7 runs, 10 loops each) # ba performance

dt = 1000
640 ms ± 9.23 ms per loop (mean ± std. dev. of 7 runs, 1 loop each) # This PR
595 ms ± 9.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each) # ba performance
```

Pinging @RonaldAJ ; could I please bother you for a verification of these results?