@larsoner 2020-11-17
On `master`:
```
In [8]: b, a = scipy.signal.iirdesign(0.2,0.3,2,80)
In [9]: sos = scipy.signal.iirdesign(0.2,0.3,2,80, output='sos')
In [10]: x = np.random.RandomState(0).randn(44100 * 100)                                                                                                                            
In [11]: %timeit scipy.signal.lfilter(b, a, x)                                                                                                                                      
30.8 ms ± 639 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
In [12]: %timeit scipy.signal.sosfilt(sos, x)                                                                                                                                       
38.2 ms ± 236 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
And on this PR:
```
In [15]: %timeit scipy.signal.lfilter(b, a, x)                                                                                                                                      
31.1 ms ± 250 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
In [16]: %timeit scipy.signal.sosfilt(sos, x)                                                                                                                                       
32 ms ± 163 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
CIs are happy and I think we can trust existing tests. Thanks @ilayn !