@RonaldAJ 2020-11-24
Using the script from issue #12970 with an extra line for printing the scipy version I have:

scipy.__version__: 1.6.0.dev0+06f51e3
time elapsed  for ba-filter: 0.06401501893997193
time elapsed for sos-filter: 0.056847667694091795

So indeed performance is now comparable for these settings.

I found however that performance for shorter signals is considerably better (as in shorter computation time per sample). I will dive into it a bit more and make a new issue for it.