@RonaldAJ 2020-11-24
For true comparison with the previous scipy version built on the same system:
scipy.__version__: 1.6.0.dev0+9d60a26
time elapsed  for ba-filter: 0.04203344583511352
time elapsed for sos-filter: 0.05560921430587769